+++ code/powertools/netcore/SampleApp/SampleApp.sln
@@ -7,8 +7,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{DEAF7F92-411
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SampleApp", "src\SampleApp\SampleApp.csproj", "{FC5B185F-DEBF-46D8-B864-9EF757827789}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LambdaHandlers", "src\LambdaHandlers\LambdaHandlers.csproj", "{469507B7-8C72-4995-81C9-64C7D8F2B593}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -26,14 +24,9 @@ Global
 		{2A90F257-DC71-4AF0-AF1D-4B557D4CD4B7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2A90F257-DC71-4AF0-AF1D-4B557D4CD4B7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2A90F257-DC71-4AF0-AF1D-4B557D4CD4B7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{469507B7-8C72-4995-81C9-64C7D8F2B593}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{469507B7-8C72-4995-81C9-64C7D8F2B593}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{469507B7-8C72-4995-81C9-64C7D8F2B593}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{469507B7-8C72-4995-81C9-64C7D8F2B593}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{FC5B185F-DEBF-46D8-B864-9EF757827789} = {DEAF7F92-411E-4FE7-823B-C4F5DB41BBF8}
 		{2A90F257-DC71-4AF0-AF1D-4B557D4CD4B7} = {DEAF7F92-411E-4FE7-823B-C4F5DB41BBF8}
-		{469507B7-8C72-4995-81C9-64C7D8F2B593} = {DEAF7F92-411E-4FE7-823B-C4F5DB41BBF8}
 	EndGlobalSection
 EndGlobal
