+++ code/powertools/netcore/SampleApp/src/SampleApp/LambdaHandlers/GetAllItemsFunction.cs
@@ -1,17 +1,13 @@
-using System;
 using System.Net;
 using System.Text.Json;
-using System.Threading.Tasks;
 using Amazon.Lambda.APIGatewayEvents;
 using Amazon.Lambda.Core;
-using Microsoft.Extensions.DependencyInjection;
 using SampleApp.Repositories;
 
 // Assembly attribute to enable the Lambda function's JSON input to be converted into a .NET class.
 [assembly: LambdaSerializer(typeof(Amazon.Lambda.Serialization.SystemTextJson.DefaultLambdaJsonSerializer))]
-namespace LambdaHandlers
+namespace SampleApp.LambdaHandlers
 {
-
     public class GetAllItemsFunction
     {
         readonly ServiceProvider _serviceProvider;
