+++ code/powertools/netcore/SampleApp/src/SampleApp/LambdaHandlers/GetItemByIDFunction.cs
@@ -1,14 +1,11 @@
 using System.Net;
 using System.Text.Json;
-using System.Threading.Tasks;
 using Amazon.Lambda.APIGatewayEvents;
 using Amazon.Lambda.Core;
-using Microsoft.Extensions.DependencyInjection;
 using SampleApp.Repositories;
 
-namespace LambdaHandlers
+namespace SampleApp.LambdaHandlers
 {
-
     public class GetItemByIDFunction
     {
         readonly ServiceProvider _serviceProvider;
@@ -17,7 +14,7 @@ namespace LambdaHandlers
         {
             Common.Initialize(out _serviceProvider);
         }
-
+        // GET api/items/{id}
         public async Task<APIGatewayProxyResponse> FunctionHandler(APIGatewayProxyRequest apigProxyEvent, ILambdaContext context)
         {
 
@@ -29,7 +26,7 @@ namespace LambdaHandlers
 
                 var result = await itemRepository.GetByIdAsync(int.Parse(id));
 
-                return (result == null) ? Common.APIResponse(HttpStatusCode.NotFound, null) : Common.APIResponse(HttpStatusCode.OK, JsonSerializer.Serialize(result));
+                return (result == null) ? Common.APIResponse(HttpStatusCode.NotFound, string.Empty) : Common.APIResponse(HttpStatusCode.OK, JsonSerializer.Serialize(result));
             }
         }
     }
