+++ code/powertools/netcore/SampleApp/src/SampleApp/SampleApp.csproj
@@ -17,6 +17,7 @@
   <ItemGroup>
     <PackageReference Include="Amazon.Lambda.AspNetCoreServer.Hosting" Version="1.6.0" />
     <PackageReference Include="AWSSDK.DynamoDBv2" Version="3.7.203.1" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="7.0.0" />
   </ItemGroup>
 
   <ItemGroup>
@@ -25,5 +26,9 @@
 
   <ItemGroup>
     <None Remove="Properties\launchSettings.json" />
+    <None Remove="LambdaHandlers\" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="LambdaHandlers\" />
   </ItemGroup>
 </Project>
